app: Created procedure call tokens returned by system/app/launch
NOTE: As system/app/launch is free for anyone to call, this doesn't
provide any additional security or validation for the listed
procedures (as client-side JS can be trivially intercepted).

Call tokens DO provide a mechanism to easily prohibit direct calls
to any procedure not listed in an app manifest. This makes it more
explicit which procedures must be hardened. It is also easy to keep
everything in sync, as allowed procedures need only be declared
once in the app manifest.