wip: Created procedure call tokens returned by system/app/launch
NOTE: As system/app/launch is free for anyone to call, this doesn't
actually provide any additional security at all. It is trivial to
call this proc and use the tokens in manually created calls instead.

A proper trust chain needs to be established from the initial page
load and provided in each system/app/launch call in some way that
isn't trivial to fake. Possibly using a shared random number
generator or similar mechanism?